Split up type-utilities.ts
