+++ src/language/semantics/type-system.ts
@@ -1,3 +1,4 @@
+export { literalTypeFromSemanticGraph } from './type-system/literal-type.js'
 export * as types from './type-system/prelude-types.js'
 export { isAssignable } from './type-system/subtyping.js'
 export {
@@ -16,19 +17,22 @@ export {
   resolveParameterTypes,
 } from './type-system/type-inference.js'
 export {
-  applicableFunctionSignatures,
-  applyKeyPathToType,
-  containedTypeParameters,
   functionParameterKey,
   functionReturnKey,
-  getTypesForTypeParameters,
-  literalTypeFromSemanticGraph,
-  replaceAllTypeParametersWithTheirConstraints,
   stringifyTypeKeyPathForEndUser,
-  supplyTypeArgument,
-  supplyTypeArguments,
   typeKeyPathFromObjectNode,
   typeParameterAssignableToConstraintKey,
-  typeParameterIdentitiesWithinType,
   type TypeKeyPath,
-} from './type-system/type-utilities.js'
+} from './type-system/type-key-path.js'
+export {
+  containedTypeParameters,
+  typeParameterIdentitiesWithinType,
+} from './type-system/type-parameter-analysis.js'
+export {
+  applicableFunctionSignatures,
+  applyKeyPathToType,
+  getTypesForTypeParameters,
+  replaceAllTypeParametersWithTheirConstraints,
+  supplyTypeArgument,
+  supplyTypeArguments,
+} from './type-system/type-substitution.js'
